reorganize local/storage layout and split distributed headers
