+++ src/umbp/distributed/peer/peer_service.cpp
@@ -28,8 +28,8 @@
 #include <cstring>
 
 #include "mori/utils/mori_log.hpp"
-#include "umbp/common/types.h"
 #include "umbp/distributed/pool_client.h"
+#include "umbp/distributed/types.h"
 #include "umbp/local/block_index/local_block_index.h"
 #include "umbp/local/storage/local_storage_manager.h"
 #include "umbp_peer.grpc.pb.h"
