+++ src/umbp/include/umbp/distributed/master/global_block_index.h
@@ -28,7 +28,7 @@
 #include <utility>
 #include <vector>
 
-#include "umbp/common/types.h"
+#include "umbp/distributed/types.h"
 
 namespace mori::umbp {
