+++ src/umbp/include/umbp/distributed/routing/route_put_strategy.h
@@ -26,7 +26,7 @@
 #include <string>
 #include <vector>
 
-#include "umbp/common/types.h"
+#include "umbp/distributed/types.h"
 
 namespace mori::umbp {
