+++ src/umbp/include/umbp/local/storage/segment/segment_scanner.h
@@ -23,8 +23,8 @@
 
 #include <string>
 
-#include "umbp/io/storage_io_driver.h"
 #include "umbp/local/storage/segment/segment_index.h"
+#include "umbp/storage/io/storage_io_driver.h"
 
 namespace mori::umbp::segment {
