+++ src/umbp/include/umbp/local/storage/segment/segment_writer.h
@@ -25,9 +25,9 @@
 #include <string>
 #include <vector>
 
-#include "umbp/io/storage_io_driver.h"
 #include "umbp/local/storage/segment/segment_format.h"
 #include "umbp/local/storage/segment/segment_index.h"
+#include "umbp/storage/io/storage_io_driver.h"
 
 namespace mori::umbp::segment {
