+++ tests/cpp/umbp/pool/test_types.cpp
@@ -21,7 +21,7 @@
 // SOFTWARE.
 #include <gtest/gtest.h>
 
-#include "umbp/common/types.h"
+#include "umbp/distributed/types.h"
 
 namespace mori::umbp {
